+++ frontend/src/components/LanguageSwitcher.vue
